removed NavLinks
